Failed Compression
Spent some time working on file compression. The libary I had planned on
using turned out to be incompatable with the .NET version I was using,
so I had to resport to using a gZipStream. I tried for a bit, but could
not get it to work. At least I now how each session within its own
directory, so that alone will help with file management.