Game Manager Update
Big update to the functionality of the game manager. All that is left is
to jumble the order of the targets, instatiate the targets, and then set
up a time delay before moving on to the next target (One frame is a
little fast).

Running into bug where the program is trying to write out to the last
binary file after it has closed. Will need to trouble shoot the logic
soon.

Also, need to design a new program to examine the session data. My
simple FloatReader.exe isn't cutting it anymore.